added coin events

+++ CcTalk/Bill/CcTalkBillEvent.cs
@@ -1,7 +1,7 @@
 namespace CcTalk.Bill;
 
-public readonly struct CcTalkBillEvent(int index, byte position) : ICcTalkBillEvent
+public readonly struct CcTalkBillEvent(byte index, byte position) : ICcTalkBillEvent
 {
-    public int Index { get; } = index;
+    public byte Index { get; } = index;
     public CcTalkBillPosition Position { get; } = (CcTalkBillPosition)position;
 }
